Develop markup, copy, and styling.
Did a very basic, breadth-first refinement of each section. A lot of HTML markup / content was added, but not necessarily with the appropriate grouping or tags.
I attempted to only address the basic "shape" of each section. I am trying this breadth-first approach for the first time, as I think it will get the site as relatively pleasant looking as fast as possible. It will also help me pinpoint commonalities between sections.

Some changes:
- All sections had their ugly, temporary background colors removed.
- Header and Realtor section has a tad more markup and starting structure.
- The Features section now have very basic cards and mroe of their markup / copy.
- The Story section(s) have their image and copy content.
- The Homes section now have very basic cards and more of their markup / copy.
- The Gallery section now has images rather than colored divs.
- The Footer section's elements now look more like links, pointer and basic hover state