Do first design-based styling pass
Up until now, content was quickly placed approximately where it should be placed, with no thought to sizing, spacing, whitespace, visual hierarchy, etc.
In this commit, I jumped around and tackled many of the concepts above, just getting the design a little more pleasant looking with the least amount of code. There are still quite a few small design decisions to be made!